[refactor] 폴더 이동
